Refactor project structure, move scripts, update docs, and add dependencies
